rename tiup-dm to tiup-dms (#429)
* rename tiup-dm to tiup-dms
