+++ components/dms/command/check.go
