+++ components/dms/command/destroy.go
