+++ components/dms/main.go
@@ -13,7 +13,7 @@
 
 package main
 
-import "github.com/pingcap/tiup/components/dm/command"
+import "github.com/pingcap/tiup/components/dms/command"
 
 func main() {
 	command.Execute()
