+++ tests/tiup-dms/test_dm_cmd.sh
